requirements.txt
streamlit
pillow
pytesseract
requests